// test/gms-soh-config-tool/src/routes/types.ts
export enum AppSections {
  STATION_GROUPS = 'Station Groups',
  GROUP = 'Group',
  STATION_GROUP_CAPABILITY = 'Station Group Capability',
  STATION_CAPABILITY = 'Station Capability',
  CHANNEL_CAPABILITY = 'Channel Capability',
}

export enum SharedCapabilityErrorTypes {
  NO_ROLLUPS = 'NO_ROLLUPS',
  MARGINAL_EXCEEDS_GOOD = 'MARGINAL_EXCEEDS_GOOD',
  THRESHOLD_EXCEEDS_MAX = 'THRESHOLD_EXCEEDS_MAX',
}

export enum StationGroupCapabilityErrorTypes {
  NO_STATIONS = 'NO_STATIONS',
  NO_ROLLUPS = 'NO_ROLLUPS',
  MARGINAL_EXCEEDS_GOOD = 'MARGINAL_EXCEEDS_GOOD',
  THRESHOLD_EXCEEDS_MAX = 'THRESHOLD_EXCEEDS_MAX',
}

export enum StationCapabilityErrorTypes {
  NO_CHANNELS = 'NO_CHANNELS',
  NO_ROLLUPS = 'NO_ROLLUPS',
  MARGINAL_EXCEEDS_GOOD = 'MARGINAL_EXCEEDS_GOOD',
  THRESHOLD_EXCEEDS_MAX = 'THRESHOLD_EXCEEDS_MAX',
}

export enum ChannelCapabilityErrorTypes {
  NO_MONITORS = 'NO_MONITORS',
  NO_ROLLUPS = 'NO_ROLLUPS',
  MARGINAL_EXCEEDS_GOOD = 'MARGINAL_EXCEEDS_GOOD',
  THRESHOLD_EXCEEDS_MAX = 'THRESHOLD_EXCEEDS_MAX',
}